Display solved puzzles on grid page
In app.py, render grid.html instead of result.html for solved puzzles.
When doing so, pass along solved.to_puzzle_string and a 'solved' boolean.
In grid.html, add a check for solved puzzles.
If the puzzle is solved, output the solution.
Otherwise, output the grid and any error messages, checking the puzzle values.
Move the validity checking to a separate checkAll function. (This function takes
    several seconds to run on a complete puzzle.)